added field validation to validators, tests added
