+++ src/common/utils/validator.js
@@ -386,6 +386,12 @@ function handleResponseError(parsedApiResponse) {
   }
 }
 
+function sanitiseValue(input, type) {
+  const regex = (type === 'year') ? '[0-9]{1,4}' : '[0-9]{1,2}';
+
+  return ((input.match(regex) === null) ? '' : input.match(regex)[0]);
+}
+
 module.exports = {
   notEmpty,
   email,
@@ -422,4 +428,5 @@ module.exports = {
   isValidRegistrationLength,
   isValidDepAndArrDate,
   handleResponseError,
+  sanitiseValue,
 };
